Add Fledgling Test to the Encap Functionality
Encapping code was written for Windows PE, Linux ELF, and generic
file formats (plain appending).  While it turned out to be much hairier
code than expected--BINARY! parsing wasn't quite all it was cracked up
to be--it did kind of work.

The code sat around and atrophied for a while, but has been brought back
to a running state.  This adds a test to show it working on Linux,
Windows, and Mac.